requirements.txt
pandas
matplotlib
ipython
ultralytics
numpy
fonttools
hanzipy
pillow
opencv-python
scikit-learn
torchmetrics
torchmetrics[detection]

--index-url https://download.pytorch.org/whl/cu126
torch
torchaudio
torchvision